aquariux_python/requirement.txt
pytest
pytest-playwright
playwright
allure-pytest
python-dotenv
pytest-html